feat: main 뼈대로직 구현

+++ src/main/java/lotto/model/PurchaseAmount.java
@@ -16,7 +16,7 @@ public class PurchaseAmount {
         }
     }
 
-    public float getPurchaseAmount() {
+    public long getPurchaseAmount() {
         return amount;
     }
